# Parallelize checking for 10x performance win