<!--
Thank you for submitting a PR to cargo-spellcheck!
-->

## What does this PR accomplish?
- Improves checking times by ~10x

comparison with `cargo-spellcheck 0.12.4` by running `cargo spellcheck` on https://github.com/rust-lang/rust/tree/22e9fe644ea710eec50cb0aabcae7fa8dd9fd675

(see below note for why I've done `-j $(nproc)`
```console
$ hyperfine "cargo spellcheck -j $(nproc)" "~/code/cargo-spellcheck/target/release/cargo-spellcheck -j $(nproc)" --export-markdown hyperfine.md 
Benchmark 1: cargo spellcheck -j 16
  Time (mean ± σ):     356.515 s ± 11.933 s    [User: 409.724 s, System: 155.652 s]
  Range (min … max):   340.755 s … 374.962 s    10 runs

Benchmark 2: ~/code/cargo-spellcheck/target/release/cargo-spellcheck -j 16
  Time (mean ± σ):     32.953 s ±  2.922 s    [User: 353.613 s, System: 13.498 s]
  Range (min … max):   26.296 s … 36.698 s    10 runs

Summary
  '~/code/cargo-spellcheck/target/release/cargo-spellcheck -j 16' ran
   10.82 ± 1.03 times faster than 'cargo spellcheck -j 16'
```

| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `cargo spellcheck -j 16` | 356.515 ± 11.933 | 340.755 | 374.962 | 10.82 ± 1.03 |
| `~/code/cargo-spellcheck/target/release/cargo-spellcheck -j 16` | 32.953 ± 2.922 | 26.296 | 36.698 | 1.00 |


<!---
Delete all that do not apply:
-->

 * 🩹 Bug Fix
 * 🦚 Feature
 
<!---
Mention the linked issue here.
This will magically close the issue once the PR is merged.
-->
Closes # .

## Changes proposed by this PR:
- Parallelise running the checkers
<!---
Tell the reviewer What changed, Why, and How were you able to accomplish that?
-->

## Notes to reviewer:
Not sure why the previous checking code was single threaded - it was definitely _attempting_ to be parallel, but misunderstood the stream APIs.

<!---
Leave a message to whoever is going to review this PR.
Mainly, pointers to review the PR, and how they can test it.
If things are still WIP or feedback on particular impl details
are wanted, state them here too.
-->

- There are probably a bunch more perf improvements to make - suspect the next bottleneck is document collection.
- Would be nice to use `tracing`, and have better (parallel) profiling tools.
- Not sure why we use the number of physical cpus - hunspell is using only half the threads it can. We should use `num_cpus::get` instead of `num_cpus::get_physical` (or `std::thread::available_parallelism`).
- clippy is _very unhappy_ with `doc-chunks`
- lots of methods on `Document` should probably be removed


## 📜 Checklist

 * [x] Works on the `./demo` sub directory
 * [x] Test coverage is excellent and passes
   No change
 * [x] Documentation is thorough
   No change