@drahnr 2023-07-14
My results appear to be very different on 40k loc project using the following command:

```
cargo b --release; hyperfine --show-output -r 2 "./target/release/cargo-spellcheck check -j16 --code 0 ../secret/project"
```

Current main branch
```
# git 8e4cfcce7f8cf3984664feba44e90dabe9d9fcf1
Benchmark 1: ./target/release/cargo-spellcheck check -j16 --code 0 ../gensyn/node
  Time (mean ± σ):     191.1 ms ±   5.7 ms    [User: 246.0 ms, System: 55.1 ms]
  Range (min … max):   187.1 ms … 195.2 ms    2 runs
```

```
# git 6c50e07efa303aa37fff73be1d7b187b36dc3e22
Benchmark 1: ./target/release/cargo-spellcheck check -j16 --code 0 ../gensyn/node
  Time (mean ± σ):     184.6 ms ±   3.2 ms    [User: 236.8 ms, System: 58.1 ms]
  Range (min … max):   182.4 ms … 186.9 ms    2 runs
```

which aligns with the CPU utilization I see - it's only one core regardless.