@drahnr 2023-07-14
```
hyperfine --show-output "./cargo-spellcheck-8e4cfcce7f8cf3984664feba44e90dabe9d9fcf1 -j 16 --code 0 ../gensyn/node" "./cargo-spellcheck-6c50e07efa303aa37fff73be1d7b187b36dc3e22 -j 16 --code 0 ../gensyn/node" --export-markdown hyperfine.md 
Benchmark 1: ./cargo-spellcheck-8e4cfcce7f8cf3984664feba44e90dabe9d9fcf1 -j 16 --code 0 ../gensyn/node
  Time (mean ± σ):     180.1 ms ±   6.3 ms    [User: 238.3 ms, System: 59.6 ms]
  Range (min … max):   170.1 ms … 191.2 ms    15 runs
 
Benchmark 2: ./cargo-spellcheck-6c50e07efa303aa37fff73be1d7b187b36dc3e22 -j 16 --code 0 ../x
  Time (mean ± σ):     178.5 ms ±   6.7 ms    [User: 235.5 ms, System: 57.5 ms]
  Range (min … max):   171.0 ms … 191.0 ms    15 runs
 
Summary
  ./cargo-spellcheck-6c50e07efa303aa37fff73be1d7b187b36dc3e22 -j 16 --code 0 ../gensyn/node ran
    1.01 ± 0.05 times faster than ./cargo-spellcheck-8e4cfcce7f8cf3984664feba44e90dabe9d9fcf1 -j 16 --code 0 ../x
    ```