@aatifsyed 2023-07-14
Interesting - just to clarify my understanding, on _this_ branch you're seeing only a single core used?