@drahnr 2023-07-15
> Interesting - just to clarify my understanding, on _this_ branch you're seeing only a single core used?

Yes, exactly. The pattern differs, but there is only ever one CPU used at 100% with this branch. I suspected it being IO bound or a lock in `stdout` (hypothesis).