@aatifsyed 2023-07-16
That's very surprising.

I've prepared tracing on two branches for you:
https://github.com/aatifsyed/cargo-spellcheck/tree/parallel-traced (my code)
https://github.com/aatifsyed/cargo-spellcheck/tree/single-thread-traced (pre-existing code)

Running each will give you a `trace-<unix_seconds>.json` that you can view at https://ui.perfetto.dev, or chome://tracing in a chromium browser.

Here are the traces on my machine (for a bare `cargo spellcheck` invocation), on https://github.com/ChainSafe/forest/:
```
$ loc
--------------------------------------------------------------------------------
 Language             Files        Lines        Blank      Comment         Code
--------------------------------------------------------------------------------
 Rust                   313        51683         5975         5028        40680
 Markdown                27         4669          823            0         3846
...
```
parallel-traced (my code)
![image](https://github.com/drahnr/cargo-spellcheck/assets/38045910/2f6d3f9f-890b-40c8-a229-1e26d50fa7e5)
you can see the multiple threads of execution
[trace-parallel.json.txt](https://github.com/drahnr/cargo-spellcheck/files/12064683/trace-parallel.json.txt)


single-thread-traced (pre-existing code)
![image](https://github.com/drahnr/cargo-spellcheck/assets/38045910/26988606-544e-47c4-8340-2a6f4a4cdfd2)
you can see that there only a single thread
[trace-single-threaded.json.txt](https://github.com/drahnr/cargo-spellcheck/files/12064684/trace-single-threaded.json.txt)