@aatifsyed 2023-07-16
>  I suspected it being IO bound or a lock in stdout (hypothesis).

I'd be surpised if this was the case!