@drahnr 2023-07-17
[parallel.json.txt](https://github.com/drahnr/cargo-spellcheck/files/12067145/parallel.json.txt)
[vanilla.json.txt](https://github.com/drahnr/cargo-spellcheck/files/12067146/vanilla.json.txt)

I am on fedora 38 AMD 64 with a 16 core processor.