@drahnr 2023-07-17
[forest-parallel.json.txt](https://github.com/drahnr/cargo-spellcheck/files/12067310/forest-parallel.json.txt)
[forest-vanilla.json.txt](https://github.com/drahnr/cargo-spellcheck/files/12067311/forest-vanilla.json.txt)

When running on `forest`, I indeed see the ~10x speedup.

---

I'll have to debug why this is not the case for the private repo, but I don't see any obstacles to merging this as is.