@drahnr 2023-07-17
Alright, I think I figured the root cause, it's rather unintuitive cli flag interpretation.